pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI Application with Blue-Green Deployment"
authors = [
    { name = "TaekGwan Kim", email = "sbe07032@naver.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11, <3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-decouple"
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "ipykernel",
    "langgraph-cli[inmem]>=0.1.55",
]

# Group for the minimal dependencies to run just the client and Streamlit app.
# These are also installed in the default dependencies.
# To install run: `uv sync --frozen --only-group client`
client = [
    "httpx~=0.27.2",
    "pydantic ~=2.10.1",
    "python-dotenv ~=1.0.1",
    "streamlit~=1.40.1",
]

